requirements.txt
dataframe_image==0.2.3
google_api_python_client==2.129.0
google_auth_oauthlib==1.2.0
matplotlib==3.8.3
MLB_StatsAPI==1.7.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
protobuf==5.26.1
pyarrow==15.0.2
pybaseball==2.2.7
python_dateutil==2.9.0.post0
pytz==2024.1
scikit_learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
xgboost==2.0.3
gunicorn>20.0.0
functions-framework==3.*
psycopg2-binary==2.9.9
sqlalchemy==2.0.*
greenlet==3.0.3